+++ src/main/java/manage/store/inventory/dto/ItemUpdateDTO.java
@@ -1,9 +1,11 @@
 package manage.store.inventory.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class ItemUpdateDTO {
 
-    private Integer quantity;
+    private BigDecimal quantity;
 }
